+++ vtm/src/org/oscim/layers/LocationLayer.java
@@ -48,7 +48,7 @@ public class LocationLayer extends Layer {
     public void setPosition(double latitude, double longitude, double accuracy) {
         double x = MercatorProjection.longitudeToX(longitude);
         double y = MercatorProjection.latitudeToY(latitude);
-        double radius = accuracy / MercatorProjection.groundResolution(latitude, 1);
+        double radius = accuracy / MercatorProjection.groundResolutionWithScale(latitude, 1);
         locationRenderer.setLocation(x, y, radius);
         locationRenderer.animate(true);
     }